requirements.txt
# DermaOps Dependencies
# ====================

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Workflow Orchestration
prefect>=2.14.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torcheval>=0.0.7

# Machine Learning & Data Science
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0

# Model Validation & Testing
deepchecks>=0.18.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Experiment Tracking
mlflow>=2.9.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data Download
kagglehub>=0.3.0

# API & UI
streamlit>=1.28.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
httpx>=0.25.0
